backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.9.0
